Rename types

+++ test/local-init.c
@@ -1,6 +1,6 @@
 f() void {
 	auto s struct (
 		c [8]char,
-		i [3]long long,
+		i [3]i64,
 	) = ('a');
 }
